# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API Eatinghistory",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/eatinghistorys": {
            "get": {
                "description": "list eatinghistory entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List eatinghistory entities",
                "operationId": "list-eatinghistory",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Eatinghistory"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create eatinghistory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create eatinghistory",
                "operationId": "create-eatinghistory",
                "parameters": [
                    {
                        "description": "Eatinghistory entity",
                        "name": "eatinghistory",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.Eatinghistory"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Eatinghistory"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/eatinghistorys/{id}": {
            "delete": {
                "description": "get eatinghistory by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a eatinghistory entity by ID",
                "operationId": "delete-eatinghistory",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Eatinghistory ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.Eatinghistory": {
            "type": "object",
            "properties": {
                "addedTime": {
                    "type": "string"
                },
                "foodmenuID": {
                    "type": "integer"
                },
                "mealplanID": {
                    "type": "integer"
                },
                "tasteID": {
                    "type": "integer"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
...
        "ent.Eatinghistory": {
            "type": "object",
            "properties": {
                "added_time": {
                    "description": "AddedTime holds the value of the \"added_time\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the EatinghistoryQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.EatinghistoryEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.EatinghistoryEdges": {
            "type": "object",
            "properties": {
                "foodmenu": {
                    "description": "Foodmenu holds the value of the foodmenu edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Foodmenu"
                },
                "mealplan": {
                    "description": "Mealplan holds the value of the mealplan edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Mealplan"
                },
                "taste": {
                    "description": "Taste holds the value of the taste edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Taste"
                },
                "user": {
                    "description": "User holds the value of the user edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.User"
                }
            }
        },
        "ent.Foodmenu": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the FoodmenuQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.FoodmenuEdges"
                },
                "foodmenu_name": {
                    "description": "FoodmenuName holds the value of the \"foodmenu_name\" field.",
                    "type": "string"
                },
                "foodmenu_type": {
                    "description": "FoodmenuType holds the value of the \"foodmenu_type\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.FoodmenuEdges": {
            "type": "object",
            "properties": {
                "eatinghistory": {
                    "description": "Eatinghistory holds the value of the eatinghistory edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Eatinghistory"
                    }
                },
                "owner": {
                    "description": "Owner holds the value of the owner edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.User"
                }
            }
        },
        "ent.Mealplan": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the MealplanQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.MealplanEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "mealplan_name": {
                    "description": "MealplanName holds the value of the \"mealplan_name\" field.",
                    "type": "string"
                }
            }
        },
        "ent.MealplanEdges": {
            "type": "object",
            "properties": {
                "eatinghistory": {
                    "description": "Eatinghistory holds the value of the eatinghistory edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Eatinghistory"
                    }
                },
                "owner": {
                    "description": "Owner holds the value of the owner edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.User"
                }
            }
        },
        "ent.Taste": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the TasteQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.TasteEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "taste_name": {
                    "description": "TasteName holds the value of the \"taste_name\" field.",
                    "type": "string"
                }
            }
        },
        "ent.TasteEdges": {
            "type": "object",
            "properties": {
                "eatinghistory": {
                    "description": "Eatinghistory holds the value of the eatinghistory edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Eatinghistory"
                    }
                }
            }
        },
        "ent.User": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the UserQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.UserEdges"
                },
                "email": {
                    "description": "Email holds the value of the \"email\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                }
            }
        },
        "ent.UserEdges": {
            "type": "object",
            "properties": {
                "eatinghistory": {
                    "description": "Eatinghistory holds the value of the eatinghistory edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Eatinghistory"
                    }
                },
                "foodmenu": {
                    "description": "Foodmenu holds the value of the foodmenu edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Foodmenu"
                    }
                },
                "mealplan": {
                    "description": "Mealplan holds the value of the mealplan edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Mealplan"
                    }
                }
            }
        },
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
...
}